pyproject.toml
[project]
name = "desk-noti"
version = "0.1.0"
description = "Windows 바탕화면/다운로드 폴더 더러움 감지 및 토스트 알림"
requires-python = ">=3.11"
dependencies = [
    "winotify",
    "pyyaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
desk-noti = "monitor:run"
desk-noti-scheduler = "setup_scheduler:register"
